rename : CarInfo 파일을 carDetail 패키지로 이동
